+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] 2018-07-21
+### Added
+- Added boss unit frames with health, cast bars, and a few auras. 
+- Added arena enemy / battleground flag carrier unit frames with health, cast bars, and a few auras. 
+
 ## [1.0.1] 2018-07-21
 ### Fixed
 - Actionbutton cooldown counts should hopefully no longer be stuck at "0.0".